requirements.txt
# AWS SDK
boto3==1.34.34
botocore==1.34.34

# Encryption
cryptography==42.0.2

# Vector operations
numpy==1.26.3

# JSON/Data handling
pydantic==2.5.3
python-dotenv==1.0.1

# HTTP
requests==2.31.0

# Date/Time utilities
python-dateutil==2.8.2

# OpenSearch
opensearch-py==2.4.2
requests-aws4auth==1.2.3
